pyproject.toml
[tool.poetry]
name = "llamaindex-agents"
version = "0.1.0"
description = ""
authors = ["tikendraw <tikendraksahu1029@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
llama-index-core = ">=0.10.43"
llama-index-llms-gemini = "^0.2.0"
python-dotenv = "^1.0.1"
llama-index-utils-workflow = "^0.1.0"
openinference-instrumentation-llama-index = ">=2.2.2"
opentelemetry-proto = ">=1.12.0"
opentelemetry-exporter-otlp = "^1.26.0"
opentelemetry-sdk = "^1.26.0"
llama-index-embeddings-gemini = "^0.1.8"
llama-index-readers-file = "^0.1.32"
llama-hub = "^0.0.79.post1"
llama-index-embeddings-huggingface = "^0.2.3"
flagembedding = "^1.2.11"
peft = "^0.12.0"
sentence-transformers = "^3.0.1"
llama-index-postprocessor-flag-embedding-reranker = "^0.1.3"
llama-index-vector-stores-qdrant = "^0.2.15"
printt = "^0.1.2"
llama-index-extractors-entity = "^0.1.3"
llama-index-llms-ollama = "^0.2.2"
llama-index-vector-stores-pinecone = "^0.1.9"
llama-index-vector-stores-chroma = "^0.1.10"
llama-index-llms-groq = "^0.1.4"
soundfile = "^0.12.1"
llama-index-embeddings-ollama = "^0.2.0"
qdrant-client = "^1.11.0"
matplotlib = "^3.9.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
